A62
I was able to finish the function by using a for loop to act as an exponent would and returned the answer casted as a double. For the next function, I did almost the same but made the n positive in the for loop and then divided 1 by the result of the for loop which I casted into a double.